Phase 5.5 — accept decimal unit_price + payment amount in API
Forgotten leftover from the minor-units → decimal migration.
Backend zod schemas still required line_items.unit_price and
payments.amount to be integers, rejecting any price like 1.975
(MEIZHIYUAN paste FOB CNY).

Removed .int() from:
  api/src/routes/operations.ts: line_item.unit_price, discount_pct
  api/src/routes/payments.ts:   amount

Discount also relaxed (0.5% etc still uncommon but no reason to
forbid). Qty stays integer — partial pieces don't ship.

+++ api/src/routes/operations.ts
@@ -41,8 +41,8 @@ function derivePaymentState(total: number, paid: number, status: string): Paymen
 const lineItemSchema = z.object({
   product_id: z.string().min(1),
   qty: z.number().int().positive(),
-  unit_price: z.number().int().nonnegative(),
-  discount_pct: z.number().int().min(0).max(100).default(0),
+  unit_price: z.number().nonnegative(),
+  discount_pct: z.number().min(0).max(100).default(0),
   cartons: z.number().int().nonnegative().default(0),
   inner_boxes: z.number().int().nonnegative().default(0),
   item_description: z.string().nullable().optional(),
